chore: add legend and implement export button properly
